feat: Support i18n for documentation
This commit overhauls the project structure and build process to support
documentation in both official languages. Because of the limitations of
`mkdocs`, it is necessary to treat each language as a distinct site. The
general approach used here is to `mkdocs build` each language, then copy
over a static landing page.

  * One folder under `docs/` per language for content
  * One config file per language at the project root
  * One output folder per language at `site/<lang>`
  * Images are no longer shared, as they generally have unilingual text

Having multiple sites means `mkdocs gh-deploy` is no longer effective.
So the CI configuration was also updated to a generic GitHub action for
publication of a directory to the `gh-pages` branch.